+++ lib/pages/dashboards/dashboard_page.dart
@@ -1,6 +1,7 @@
 import 'dart:math';
 
 import 'package:auto_route/auto_route.dart';
+import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:lotti/classes/entity_definitions.dart';
@@ -9,6 +10,7 @@ import 'package:lotti/get_it.dart';
 import 'package:lotti/pages/empty_scaffold.dart';
 import 'package:lotti/services/nav_service.dart';
 import 'package:lotti/themes/theme.dart';
+import 'package:lotti/utils/consts.dart';
 import 'package:lotti/widgets/app_bar/dashboard_app_bar.dart';
 import 'package:lotti/widgets/charts/dashboard_health_chart.dart';
 import 'package:lotti/widgets/charts/dashboard_measurables_chart.dart';
@@ -145,6 +147,8 @@ class DashboardWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    void beamToNamed(String path) => context.beamToNamed(path);
+
     return SingleChildScrollView(
       child: Padding(
         padding: const EdgeInsets.all(8),
@@ -160,6 +164,7 @@ class DashboardWidget extends StatelessWidget {
                 measurement: (DashboardMeasurementItem measurement) {
                   return DashboardMeasurablesChart(
                     measurableDataTypeId: measurement.id,
+                    dashboardId: dashboardId,
                     aggregationType: measurement.aggregationType,
                     rangeStart: rangeStart,
                     rangeEnd: rangeEnd,
@@ -227,10 +232,15 @@ class DashboardWidget extends StatelessWidget {
                 IconButton(
                   icon: const Icon(Icons.dashboard_customize_outlined),
                   color: colorConfig().entryTextColor,
-                  onPressed: () {
-                    navigateNamedRoute(
-                      '/settings/dashboards/$dashboardId',
-                    );
+                  onPressed: () async {
+                    final beamerNav = await getIt<JournalDb>()
+                        .getConfigFlag(enableBeamerNavFlag);
+
+                    if (beamerNav) {
+                      beamToNamed('/settings/dashboards/$dashboardId');
+                    } else {
+                      navigateNamedRoute('/settings/dashboards/$dashboardId');
+                    }
                   },
                 ),
               ],